Fix lexical scope bug in closures
Because closures were capturing the mutable enclosing environment (and
not a snapshot of the environment when the function were being declared)
it was essentially resolving variables as if using dynamic scope.

To fix this, we added a Resolver which figures out for each variable
expression the number of hops in the environment chain needed to reach the
declaring scope of the variable in question.

Another potential solution would have been to implement snapshot
environments, which are essentially immutable environments which are
used as the enclosing environment of closures. However, the Resolver
solution had the advantage of also speeding up variable
acccess (essentially removing several `if` condition instructions on
each access.)