Improve comments and function names in compiler.c
compiler.c is probably the hardest piece in this project due to the fact
that it conflates parsing and compilation, but also because parsing is
done using a Pratt parser.

Pratt parsers are sleek (since they are essentially a highly refactored form
of recursive descent parsers) but they can also be hard to understand
unless they are well documented.

This commit attempts to address both concerns above.