Write a proper Makefile
The major improvement is that we now don't have to manually track
dependencies for each source file. This is now automatically performed
as a byproduct of compilation.

The Makefile is now harder to understand, but this is a fine price to
pay for the benefit of getting reliable builds.

We also changed the layout of the project a bit (see the Makefile for details)