Report as many useful errors as possible
We use the panic mode flag to detect when a parsing rule was violated,
and attempt to synchronize to a subsequent part of the input that could
be correctly parsed.

However, as soon as an error is found, the whole compilation is
considered compromise, so the resulting (incomplete) code won't be
executed.